Some structural reshuffling and renaming of directories for clarity
