+++ .gitignore
@@ -223,4 +223,4 @@ project.lock.json
 #####
 # End of core ignore list, below put you custom 'per project' settings (patterns or path)
 #####
-/TextMUD/Jsons/
+/TextMUD/FileIO/Jsons/
