+++ TextMUD/FileIO/SaveGameHandle/SaveLoader.cs
@@ -9,7 +9,7 @@ namespace TextMUD.FileIO.SaveGameHandle
     {
         public static Eukaryote Load(string name)
         {
-            string path = @"C:\Users\Peter\RiderProjects\TextMUD\TextMUD\Jsons\" + $"{name}.json";
+            string path = @"C:\Users\Peter\RiderProjects\TextMUD\TextMUD\FileIO\Jsons\" + $"{name}.json";
             //get data from path
             string data = File.ReadAllText(path);
             //return as deserialized object
