+++ TextMUD/Terrain/Room.cs
@@ -2,6 +2,6 @@
 {
     public class Room
     {
-        
+        //TODO imp, "Room"
     }
 }
